+++ AddOns/Platynator/Core/Migration.lua
@@ -650,6 +650,35 @@ local function UpgradeDesignv15(design)
   end
 end
 
+local function UpgradeDesignv16(design)
+  local function UpdateAutoColorsv16(autoColors)
+    for _, ac in ipairs(autoColors) do
+      if ac.kind == "threatIgnoreRole" then
+        ac.colors.offtank = GetColor("0FAAC8")
+        ac.useOffTankColor = false
+      end
+    end
+  end
+
+  for _, b in ipairs(design.bars) do
+    if b.autoColors then
+      UpdateAutoColorsv16(b.autoColors)
+    end
+  end
+
+  for _, h in ipairs(design.highlights) do
+    if h.autoColors then
+      UpdateAutoColorsv16(h.autoColors)
+    end
+  end
+
+  for _, t in ipairs(design.texts) do
+    if t.autoColors then
+      UpdateAutoColorsv16(t.autoColors)
+    end
+  end
+end
+
 local designUpgrades = {
   UpgradeDesignv1,
   UpgradeDesignv2,
@@ -666,6 +695,7 @@ local designUpgrades = {
   UpgradeDesignv13,
   UpgradeDesignv14,
   UpgradeDesignv15,
+  UpgradeDesignv16,
 }
 
 function addonTable.Core.UpgradeDesign(design)
@@ -815,35 +845,22 @@ local function MigrateSettingsv6()
   addonTable.Display.Utilities.MigrateAuraFilters()
 end
 
+local settingUpgrades = {
+  MigrateSettingsv1,
+  MigrateSettingsv2,
+  MigrateSettingsv3,
+  MigrateSettingsv4,
+  MigrateSettingsv5,
+  MigrateSettingsv6,
+  MigrateSettingsv6,
+}
 function addonTable.Core.MigrateSettings()
-  if addonTable.Config.Get(addonTable.Config.Options.MIGRATION) == 1 then
-    MigrateSettingsv1()
-    addonTable.Config.Set(addonTable.Config.Options.MIGRATION, 2)
-  end
+  if #settingUpgrades + 1 ~= addonTable.Config.Get(addonTable.Config.Options.MIGRATION) then
+    for i = addonTable.Config.Get(addonTable.Config.Options.MIGRATION), #settingUpgrades do
+      settingUpgrades[i]()
+    end
 
-  if addonTable.Config.Get(addonTable.Config.Options.MIGRATION) == 2 then
-    MigrateSettingsv2()
-    addonTable.Config.Set(addonTable.Config.Options.MIGRATION, 3)
-  end
-
-  if addonTable.Config.Get(addonTable.Config.Options.MIGRATION) == 3 then
-    MigrateSettingsv3()
-    addonTable.Config.Set(addonTable.Config.Options.MIGRATION, 4)
-  end
-
-  if addonTable.Config.Get(addonTable.Config.Options.MIGRATION) == 4 then
-    MigrateSettingsv4()
-    addonTable.Config.Set(addonTable.Config.Options.MIGRATION, 5)
-  end
-
-  if addonTable.Config.Get(addonTable.Config.Options.MIGRATION) == 5 then
-    MigrateSettingsv5()
-    addonTable.Config.Set(addonTable.Config.Options.MIGRATION, 6)
-  end
-
-  if addonTable.Config.Get(addonTable.Config.Options.MIGRATION) == 6 then
-    MigrateSettingsv6()
-    addonTable.Config.Set(addonTable.Config.Options.MIGRATION, 7)
+    addonTable.Config.Set(addonTable.Config.Options.MIGRATION, #settingUpgrades + 1)
   end
 
   for _, design in pairs(addonTable.Config.Get(addonTable.Config.Options.DESIGNS)) do