+++ AddOns/Platynator/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Platynator
 
-## [454](https://github.com/TheMouseNest/Platynator/tree/454) (2026-08-09)
-[Full Changelog](https://github.com/TheMouseNest/Platynator/compare/453...454) 
+## [458](https://github.com/TheMouseNest/Platynator/tree/458) (2026-08-13)
+[Full Changelog](https://github.com/TheMouseNest/Platynator/compare/457...458) 
 
-- Fix settings migration error on import  
+- Fix error in customise dialog in some designs  
+- Fix migration steps for adding aura filters for current spec id  
+- Ignore buffs on Valeera specifically due to showing too many  
+- Include explicit party role assignments when detecting off-tank  
+- Auto colors: Add off-tank to "Threat (no role swap)"